refactor: clean up directory
- 컴포넌트의 스타일과 스토리북을 컴포넌트 폴더 안에 추가
- 'Page' postfix 추가
- 컴포넌트 파일명을 index로 변경
- module css 파일명을 index.module로 변경
