+++ src/components/@common/Modal/index.jsx
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import styles from 'css/module/Modal.module.css';
+import styles from './index.module.css';
 
 const Modal = ({ handleModal, children }) => {
   return (
