+++ src/components/CardPreview/index.jsx
