+++ src/page/CardAddPage/index.jsx
@@ -2,12 +2,12 @@ import { useReducer, useCallback, useRef } from 'react';
 import useCardInfo from 'hooks/useCardInfo';
 import { useNavigate } from 'react-router-dom';
 
-import FormInput from 'components/common/FormInput';
+import FormInput from 'components/@common/FormInput';
 import CardPreview from 'components/CardPreview';
-import Modal from 'components/common/Modal';
+import Modal from 'components/@common/Modal';
 import CardCompany from 'components/CardCompany';
-import Button from 'components/common/Button';
-import Tooltip from 'components/common/Tooltip';
+import Button from 'components/@common/Button';
+import Tooltip from 'components/@common/Tooltip';
 
 import { moveInputFocus } from 'utils';
 import { validator } from './validator';
