+++ src/page/CardAddPage/validator.js
