+++ src/page/ConfirmationPage/validator.js
